pyproject.toml
[project]
name = "llm-rag-practices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.30",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.28",
    "langchain-tavily==0.1.5",
    "langchainhub>=0.1.21",
    "langgraph>=0.6.3",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
    "tiktoken>=0.9.0",
]
